requirements.txt
Flask==2.0.1
requests>=2.28.1
pandas==1.3.3
openai>=1.0.0
pytest==6.2.4
loguru==0.5.3
python-dotenv>=0.21.0
tqdm>=4.64.1
# 添加阿里云模型API相关依赖
dashscope>=1.10.0
alibabacloud-dashscope>=0.0.2
retry>=0.9.2
# 添加模型管理脚本所需依赖
colorama>=0.4.4